exd_streams v0.0.2 ExdStreams.Store.KeyValueStore behaviour

A behaviour for key-value based state stores

Callbacks

all()

all() :: {:ok, Stream.t()}

Return an iterator over all keys in this store.

delete(key)

delete(key()) :: :ok | {:error, term()}

Delete the value from the store if one is present

get(key)

get(key()) :: {:ok, value()}

Get the value corresponding to this key.

put(key, value)

put(key(), value()) :: :ok | {:error, term()}

Update the value associated with this key

put_all(list)

put_all([{key(), value()}]) :: :ok | {:error, term()}

Update all the given key/value pairs